Runsheet-backend/requirements.txt
fastapi==0.115.0
uvicorn==0.24.0
pydantic==2.10.0
pydantic-settings==2.7.0
python-multipart==0.0.18
elasticsearch==8.11.0
python-dotenv==1.0.0
litellm==1.55.0
strands-agents==1.8.0
strands-agents-tools==0.2.7
google-generativeai==0.8.0
google-cloud-aiplatform==1.72.0

# Redis for session storage (Requirement 8.4)
redis==5.2.0

# Rate limiting (Requirements 14.1, 14.2)
slowapi==0.1.9

# OpenTelemetry for distributed tracing (Requirement 5.3)
opentelemetry-api==1.28.0
opentelemetry-sdk==1.28.0
opentelemetry-exporter-otlp==1.28.0

# Testing infrastructure (Requirement 11.6)
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Property-based testing (Requirement 11.1)
hypothesis==6.115.0

# JWT token verification (Requirement 9.6)
python-jose[cryptography]==3.3.0
